Store NULL in match_data->subject when subject was NULL.
Previously, match_data->subject would store a pointer to a stack-allocated string,
but that pointer would be dangling as soon as pcre2_match/pcre2_dfa_match returned.
These pointers were than memcpy'd (with a size of 0), which is technically undefined behaviour,
despite not causing any observable problems, even with valgrind.